Add support to run tool in lambda
It *should* autodetect (via the presence of a lambda environment variable) if
it is running in the commandline or in a lambda function or not, and proceed
accordingly.

This is intended to allow the code to be used in a more automated fashion to
monitor changes as opposed to just spot checks.